refactor project registration
