+++ src/modules/project/components/registration/editor/forms/ProjectRegistrationForm.tsx
@@ -24,8 +24,8 @@ import withWidth, { WithWidth } from '@material-ui/core/withWidth';
 import DeleteForeverIcon from '@material-ui/icons/DeleteForever';
 import PersonIcon from '@material-ui/icons/Person';
 import { IProjectDetail, IProjectSales } from '@project/classes/response';
-import DocumentForm from '@project/components/registration/forms/DocumentForm';
-import { InformationForm } from '@project/components/registration/forms/InformationForm';
+import DocumentForm from '@project/components/registration/editor/forms/DocumentForm';
+import { InformationForm } from '@project/components/registration/editor/forms/InformationForm';
 import styles from '@styles';
 import * as classNames from 'classnames';
 import * as React from 'react';
